Move bracket structs to a separate file
